Revert "Py 2.4 fixes."
PyDict_Next went from int* to Py_ssize_t*.

This reverts commit 7932ecf3e9dee5c90f2f20e8c88deab411ced3b5.

+++ Cython/Compiler/ModuleNode.py
@@ -523,9 +523,9 @@ class ModuleNode(Nodes.Node, Nodes.BlockNode):
 #endif
 
 #if PY_VERSION_HEX < 0x02050000
-  typedef long Py_ssize_t;
-  #define PY_SSIZE_T_MAX LONG_MAX
-  #define PY_SSIZE_T_MIN LONG_MIN
+  typedef int Py_ssize_t;
+  #define PY_SSIZE_T_MAX INT_MAX
+  #define PY_SSIZE_T_MIN INT_MIN
   #define PY_FORMAT_SIZE_T \"\"
   #define PyInt_FromSsize_t(z) PyInt_FromLong(z)
   #define PyInt_AsSsize_t(o)   PyInt_AsLong(o)
